requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
torch>=1.10.0
transformers>=4.20.0
tqdm>=4.62.0
Pillow>=8.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
openai>=1.0.0